♻️ refactor: 카테고리 분리 리팩터링
- Question에 종속된 구조에서 독립 구조로 분리
- Admin 폴더와의 일치성 유지
