+++ apps/qna/docs/api_descriptions.py
@@ -3,6 +3,23 @@ class ApiDescriptions:
     QnA API 문서화를 위한 상세 설명
     """
 
+    # --- Category ---
+    CATEGORY_LIST = """
+        ### 등록된 카테고리 목록을 트리구조 데이터로 조회합니다.
+        **카테고리 목록에서 조회가능한 항목**
+        - 카테고리 ID
+        - 카테고리명
+        - 카테고리 분류 타입 (대분류)
+        - 하위 카테고리
+            - 카테고리 ID
+            - 하위 카테고리의 명칭
+            - 카테고리 분류 타입 (중분류)
+            - 하위 카테고리
+                - 카테고리 ID
+                - 하위 카테고리의 명칭
+                - 카테고리 분류 타입 (소분류)
+        """
+
     # --- Questions ---
     QUESTION_CREATE = """
     ### 수강생 권한을 가진 로그인 유저가 질문을 등록합니다.
@@ -63,22 +80,6 @@ class ApiDescriptions:
                 - 유저 닉네임
             - 댓글 내용
             - 댓글 작성일시
-    """
-
-    QUESTION_CATEGORY_LIST = """
-    ### 등록된 카테고리 목록을 트리구조 데이터로 조회합니다.
-    **카테고리 목록에서 조회가능한 항목**
-    - 카테고리 ID
-    - 카테고리명
-    - 카테고리 분류 타입 (대분류)
-    - 하위 카테고리
-        - 카테고리 ID
-        - 하위 카테고리의 명칭
-        - 카테고리 분류 타입 (중분류)
-        - 하위 카테고리
-            - 카테고리 ID
-            - 하위 카테고리의 명칭
-            - 카테고리 분류 타입 (소분류)
     """
 
     # --- Answers ---
