+++ apps/qna/docs/api_response_examples.py
@@ -8,6 +8,39 @@ class SuccessResponseExamples:
     성공 응답(200, 201)에 대한 예시 데이터 모음
     """
 
+    # --- Category ---
+    CATEGORY_LIST = OpenApiExample(
+        name="카테고리 목록 조회 성공 response body 예시",
+        value={
+            "categories": [
+                {
+                    "id": 1,
+                    "name": "백엔드",
+                    "depth": 0,
+                    "subcategories": [
+                        {
+                            "id": 5,
+                            "name": "Django",
+                            "depth": 1,
+                            "subcategories": [
+                                {"id": 12, "name": "ORM", "depth": 2, "subcategories": []},
+                                {"id": 13, "name": "DRF", "depth": 2, "subcategories": []},
+                            ],
+                        },
+                        {"id": 6, "name": "Python", "depth": 1, "subcategories": []},
+                    ],
+                },
+                {
+                    "id": 2,
+                    "name": "프론트엔드",
+                    "depth": 0,
+                    "subcategories": [{"id": 8, "name": "React", "depth": 1, "subcategories": []}],
+                },
+            ]
+        },
+        response_only=True,
+    )
+
     # --- Questions ---
     QUESTION_CREATE = OpenApiExample(
         name="질문 등록 성공 response body 예시",
@@ -38,38 +71,6 @@ class SuccessResponseExamples:
                     "thumbnail_img_url": "https://cdn.ozcodingschool.com/qna/thumb_10501_01.png",
                 }
             ],
-        },
-        response_only=True,
-    )
-
-    QUESTION_CATEGORY_LIST = OpenApiExample(
-        name="카테고리 목록 조회 성공 response body 예시",
-        value={
-            "categories": [
-                {
-                    "id": 1,
-                    "name": "백엔드",
-                    "depth": 0,
-                    "subcategories": [
-                        {
-                            "id": 5,
-                            "name": "Django",
-                            "depth": 1,
-                            "subcategories": [
-                                {"id": 12, "name": "ORM", "depth": 2, "subcategories": []},
-                                {"id": 13, "name": "DRF", "depth": 2, "subcategories": []},
-                            ],
-                        },
-                        {"id": 6, "name": "Python", "depth": 1, "subcategories": []},
-                    ],
-                },
-                {
-                    "id": 2,
-                    "name": "프론트엔드",
-                    "depth": 0,
-                    "subcategories": [{"id": 8, "name": "React", "depth": 1, "subcategories": []}],
-                },
-            ]
         },
         response_only=True,
     )
@@ -159,7 +160,7 @@ class SuccessResponseExamples:
         response_only=True,
     )
 
-    # --- Admin ---
+    # --- Admin Category ---
     ADMIN_CATEGORY_CREATE = OpenApiExample(
         name="어드민 카테고리 등록 성공 response body 예시",
         value={
@@ -202,6 +203,7 @@ class SuccessResponseExamples:
         response_only=True,
     )
 
+    # --- Admin Question ---
     ADMIN_QUESTION_LIST = OpenApiExample(
         name="어드민 질의응답 목록 조회 성공 response body 예시",
         value={
@@ -266,6 +268,7 @@ class SuccessResponseExamples:
         response_only=True,
     )
 
+    # --- Admin Answer ---
     ADMIN_ANSWER_DELETE = OpenApiExample(
         name="어드민 답변 삭제 성공 response body 예시",
         value={"answer_id": 801, "deleted_comment_count": 9},
@@ -277,6 +280,13 @@ class ErrorResponseExamples:
     """
     에러 응답(400, 401, 403, 404, 409)에 대한 예시 데이터 모음
     """
+
+    # --- CATEGORY_LIST ---
+    CATEGORY_LIST_400 = OpenApiExample(
+        name="카테고리 목록 조회 실패 response body 예시 - 잘못된 요청",
+        value={"error_detail": ErrorMessages.INVALID_CATEGORY_LIST.value},
+        response_only=True,
+    )
 
     # --- QUESTION_CREATE ---
     QUESTION_CREATE_400 = OpenApiExample(
@@ -306,12 +316,7 @@ class ErrorResponseExamples:
         value={"error_detail": ErrorMessages.NOT_FOUND_QUESTION_LIST.value},
         response_only=True,
     )
-    #  ---QUESTION_CATEGORY_LIST ---
-    QUESTION_CATEGORY_LIST_400 = OpenApiExample(
-        name="카테고리 목록 조회 실패 response body 예시 - 잘못된 요청",
-        value={"error_detail": ErrorMessages.INVALID_QUESTION_CATEGORY_LIST.value},
-        response_only=True,
-    )
+
     # --- QUESTION_DETAIL ---
     QUESTION_DETAIL_400 = OpenApiExample(
         name="질문 상세 조회 실패 response body 예시 - 잘못된 요청",
