+++ apps/qna/views/category_views.py
@@ -1,0 +1,46 @@
+from drf_spectacular.utils import OpenApiResponse, extend_schema
+
+from rest_framework.permissions import AllowAny
+from rest_framework.request import Request
+from rest_framework.response import Response
+
+from apps.qna.docs.api_descriptions import ApiDescriptions
+from apps.qna.docs.api_response_examples import (
+    ErrorResponseExamples,
+    SuccessResponseExamples,
+)
+from apps.qna.services.category.query import CategoryQueryService
+from apps.qna.views.base_view import QnaBaseAPIView
+from apps.qna.serializers.category.response import CategoryTreeResponseSerializer
+
+
+class CategoryTreeAPIView(QnaBaseAPIView):
+    """
+    질의응답 카테고리 전체 계층 구조 조회 API
+    """
+
+    permission_classes = [AllowAny]
+
+    # 카테고리 목록 조회
+    # [GET] /api/v1/qna/categories
+    @extend_schema(
+        summary="카테고리 계층 구조 조회",
+        description=ApiDescriptions.CATEGORY_LIST,
+        responses={
+            200: OpenApiResponse(
+                description="OK",
+                response=CategoryTreeResponseSerializer,
+                examples=[SuccessResponseExamples.CATEGORY_LIST],
+            ),
+            400: OpenApiResponse(
+                description="Bad Request", response=dict, examples=[ErrorResponseExamples.CATEGORY_LIST_400]
+            ),
+        },
+        tags=["qna"],
+    )
+    def get(self, request: Request) -> Response:
+        categories_tree = CategoryQueryService.get_category_tree()
+
+        response_serializer = CategoryTreeResponseSerializer({"categories": categories_tree})
+
+        return Response(response_serializer.data)
